// src/gen/types/assistant.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export const assistantGraphIdEnum = {
  agent: "agent",
  staking_agent: "staking_agent",
} as const;

export type AssistantGraphIdEnumKey =
  (typeof assistantGraphIdEnum)[keyof typeof assistantGraphIdEnum];

export type Assistant = {
  /**
   * @description The ID of the assistant.
   * @type string, uuid
   */
  assistant_id: string;
  /**
   * @description The ID of the graph.
   * @type string
   */
  graph_id: AssistantGraphIdEnumKey;
  /**
   * @description The assistant config.
   * @type object
   */
  config: {
    /**
     * @type array | undefined
     */
    tags?: string[];
    /**
     * @type integer | undefined
     */
    recursion_limit?: number;
    /**
     * @type object | undefined
     */
    configurable?: object;
  };
  /**
   * @description Static context added to the assistant.
   * @type object | undefined
   */
  context?: object;
  /**
   * @description The time the assistant was created.
   * @type string, date-time
   */
  created_at: string;
  /**
   * @description The last time the assistant was updated.
   * @type string, date-time
   */
  updated_at: string;
  /**
   * @description The assistant metadata.
   * @type object
   */
  metadata: object;
  /**
   * @description The version of the assistant
   * @type integer | undefined
   */
  version?: number;
  /**
   * @description The name of the assistant
   * @type string | undefined
   */
  name?: string;
  /**
   * @description The description of the assistant
   * @type string,null | undefined
   */
  description?: string | null;
};